# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {
            "name": "API Support"
        }
    },
    "paths": {
        "/enum/delete": {
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "根据id删除指定枚举",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "enum枚举"
                ],
                "summary": "根据id删除指定枚举",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "枚举id",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.ResponseResultInfo"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.EmptyInfo"
                        }
                    }
                }
            }
        },
        "/enum/details": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "获取枚举详情",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "enum枚举"
                ],
                "summary": "获取枚举详情",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "枚举id",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.ResponseResultInfo"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.EmptyInfo"
                        }
                    }
                }
            }
        },
        "/enum/getAllEnums": {
            "get": {
                "description": "获取全部数据",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "enum枚举"
                ],
                "summary": "获取全部数据",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.ResponseResultInfo"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.EmptyInfo"
                        }
                    }
                }
            }
        },
        "/enum/getEnumsByType": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "根据分类查询枚举",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "enum枚举"
                ],
                "summary": "根据分类查询枚举",
                "parameters": [
                    {
                        "type": "string",
                        "description": "枚举类型code",
                        "name": "type",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.ResponseResultInfo"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.EmptyInfo"
                        }
                    }
                }
            }
        },
        "/enum/getEnumsList": {
            "post": {
                "description": "分页获取枚举列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "enum枚举"
                ],
                "summary": "分页获取枚举列表",
                "parameters": [
                    {
                        "description": "请求对象",
                        "name": "param",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/enum.ListForm"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.ResponseResultInfo"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.EmptyInfo"
                        }
                    }
                }
            }
        },
        "/enum/save": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "增加、编辑",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "enum枚举"
                ],
                "summary": "增加、编辑",
                "parameters": [
                    {
                        "description": "请求对象",
                        "name": "param",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/enum.SaveForm"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.ResponseResultInfo"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.EmptyInfo"
                        }
                    }
                }
            }
        },
        "/file/upload": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "文件上传",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "file文件"
                ],
                "summary": "文件上传",
                "parameters": [
                    {
                        "type": "file",
                        "description": "请求对象",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.ResponseResultInfo"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.EmptyInfo"
                        }
                    }
                }
            }
        },
        "/user/details": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "获取用户详情",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user用户"
                ],
                "summary": "获取用户详情",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户id",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.ResponseResultInfo"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.EmptyInfo"
                        }
                    }
                }
            }
        },
        "/user/edit": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "编辑用户信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user用户"
                ],
                "summary": "编辑用户信息",
                "parameters": [
                    {
                        "description": "请求对象",
                        "name": "param",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/user.EditForm"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.ResponseResultInfo"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.EmptyInfo"
                        }
                    }
                }
            }
        },
        "/user/getUserList": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "获取user用户列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user用户"
                ],
                "summary": "获取user用户列表",
                "parameters": [
                    {
                        "description": "请求对象",
                        "name": "param",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/user.ListForm"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.ResponseResultInfo"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.EmptyInfo"
                        }
                    }
                }
            }
        },
        "/user/getVerificationCode": {
            "post": {
                "description": "获取验证码",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user用户"
                ],
                "summary": "获取验证码",
                "parameters": [
                    {
                        "description": "请求对象",
                        "name": "param",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/user.VerificationCodeForm"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.ResponseResultInfo"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.EmptyInfo"
                        }
                    }
                }
            }
        },
        "/user/login": {
            "post": {
                "description": "用户登陆",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user用户"
                ],
                "summary": "用户登陆",
                "parameters": [
                    {
                        "description": "请求对象",
                        "name": "param",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/user.LoginForm"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.ResponseResultInfo"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.EmptyInfo"
                        }
                    }
                }
            }
        },
        "/user/register": {
            "post": {
                "description": "用户注册",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user用户"
                ],
                "summary": "用户注册",
                "parameters": [
                    {
                        "description": "请求对象",
                        "name": "param",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/user.RegisterForm"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.ResponseResultInfo"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.EmptyInfo"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "enum.ListForm": {
            "type": "object",
            "required": [
                "pageNo",
                "pageSize"
            ],
            "properties": {
                "name": {
                    "type": "string"
                },
                "pageNo": {
                    "type": "integer",
                    "minimum": 0,
                    "example": 1
                },
                "pageSize": {
                    "type": "integer",
                    "minimum": 0,
                    "example": 10
                },
                "typeName": {
                    "type": "string"
                }
            }
        },
        "enum.SaveForm": {
            "type": "object",
            "required": [
                "code",
                "name",
                "typeCode",
                "typeName"
            ],
            "properties": {
                "code": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "parentId": {
                    "type": "string"
                },
                "typeCode": {
                    "type": "string"
                },
                "typeName": {
                    "type": "string"
                }
            }
        },
        "user.EditForm": {
            "type": "object",
            "required": [
                "id",
                "nickName"
            ],
            "properties": {
                "avatar": {
                    "type": "string"
                },
                "gender": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "nickName": {
                    "type": "string",
                    "maxLength": 40,
                    "minLength": 4
                },
                "password": {
                    "type": "string"
                },
                "phoneNumber": {
                    "type": "string"
                },
                "role": {
                    "type": "integer"
                },
                "userAccount": {
                    "type": "string"
                }
            }
        },
        "user.ListForm": {
            "type": "object",
            "required": [
                "pageNo",
                "pageSize"
            ],
            "properties": {
                "nickName": {
                    "type": "string"
                },
                "pageNo": {
                    "type": "integer",
                    "minimum": 0,
                    "example": 1
                },
                "pageSize": {
                    "type": "integer",
                    "minimum": 0,
                    "example": 10
                },
                "phoneNumber": {
                    "type": "string"
                },
                "userAccount": {
                    "type": "string"
                }
            }
        },
        "user.LoginForm": {
            "type": "object",
            "required": [
                "password",
                "userAccount"
            ],
            "properties": {
                "password": {
                    "type": "string",
                    "maxLength": 300,
                    "minLength": 6,
                    "example": "123456"
                },
                "userAccount": {
                    "type": "string",
                    "example": "admin@163.com"
                }
            }
        },
        "user.RegisterForm": {
            "type": "object",
            "required": [
                "code",
                "password",
                "userAccount"
            ],
            "properties": {
                "code": {
                    "type": "string",
                    "maxLength": 6,
                    "minLength": 6
                },
                "password": {
                    "type": "string",
                    "maxLength": 20,
                    "minLength": 6
                },
                "userAccount": {
                    "type": "string"
                }
            }
        },
        "user.VerificationCodeForm": {
            "type": "object",
            "required": [
                "email"
            ],
            "properties": {
                "email": {
                    "type": "string"
                }
            }
        },
        "utils.EmptyInfo": {
            "type": "object"
        },
        "utils.ResponseResultInfo": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "msg": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}